requirements.txt
django==3.0.8
djangorestframework==3.11.0
djangorestframework_simplejwt==4.4.0
django-filter==2.3.0
pytest==6.0.1
pytest-cov==2.10.0
pytest-django==3.9.0
factory-boy==2.12.0
freezegun==0.3.15
drf-yasg==1.17.1
psycopg2